+++ kedro/io/memory_dataset.py
@@ -3,6 +3,8 @@
 from __future__ import annotations
 
 import copy
+import pickle
+from multiprocessing.managers import SyncManager
 from typing import Any
 
 from kedro.io.core import AbstractDataset, DatasetError
@@ -136,3 +138,40 @@ def _copy_with_mode(data: Any, copy_mode: str) -> Any:
         )
 
     return copied_data
+
+
+class SharedMemoryDataset:
+    """``_SharedMemoryDataset`` is a wrapper class for a shared MemoryDataset in SyncManager.
+    It is not inherited from AbstractDataset class.
+    """
+
+    def __init__(self, manager: SyncManager):
+        """Creates a new instance of ``_SharedMemoryDataset``,
+        and creates shared memorydataset attribute.
+
+        Args:
+            manager: An instance of multiprocessing manager for shared objects.
+
+        """
+        self.shared_memory_dataset = manager.MemoryDataset()  # type: ignore
+
+    def __getattr__(self, name):
+        # This if condition prevents recursive call when deserialising
+        if name == "__setstate__":
+            raise AttributeError()
+        return getattr(self.shared_memory_dataset, name)
+
+    def save(self, data: Any):
+        """Calls save method of a shared MemoryDataset in SyncManager."""
+        try:
+            self.shared_memory_dataset.save(data)
+        except Exception as exc:
+            # Checks if the error is due to serialisation or not
+            try:
+                pickle.dumps(data)
+            except Exception as serialisation_exc:  # SKIP_IF_NO_SPARK
+                raise DatasetError(
+                    f"{str(data.__class__)} cannot be serialised. ParallelRunner "
+                    "implicit memory datasets can only be used with serialisable data"
+                ) from serialisation_exc
+            raise exc
