+++ tests/runner/test_parallel_runner.py
@@ -13,6 +13,7 @@ from kedro.io import (
     DatasetError,
     LambdaDataset,
     MemoryDataset,
+    SharedMemoryDataset,
 )
 from kedro.pipeline import node
 from kedro.pipeline.modular_pipeline import pipeline as modular_pipeline
@@ -21,7 +22,6 @@ from kedro.runner.parallel_runner import (
     _MAX_WINDOWS_WORKERS,
     ParallelRunnerManager,
     _run_node_synchronization,
-    _SharedMemoryDataset,
 )
 from tests.runner.conftest import (
     exception_fn,
@@ -54,7 +54,7 @@ class TestValidParallelRunner:
     def test_create_default_dataset(self):
         # dataset is a proxy to a dataset in another process.
         dataset = ParallelRunner().create_default_dataset("")
-        assert isinstance(dataset, _SharedMemoryDataset)
+        assert isinstance(dataset, SharedMemoryDataset)
 
     @pytest.mark.parametrize("is_async", [False, True])
     def test_parallel_run(self, is_async, fan_out_fan_in, catalog):
